adding AnnotationActions
