+++ client/app/reader/actions.js
@@ -12,105 +12,9 @@ export const collectAllTags = (documents) => ({
   payload: documents
 });
 
-export const startEditAnnotation = (annotationId) => ({
-  type: Constants.START_EDIT_ANNOTATION,
-  payload: {
-    annotationId
-  },
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'start-edit-annotation'
-    }
-  }
-});
-
-export const openAnnotationDeleteModal = (annotationId, analyticsLabel) => ({
-  type: Constants.OPEN_ANNOTATION_DELETE_MODAL,
-  payload: {
-    annotationId
-  },
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'open-annotation-delete-modal',
-      label: analyticsLabel
-    }
-  }
-});
-export const closeAnnotationDeleteModal = () => ({
-  type: Constants.CLOSE_ANNOTATION_DELETE_MODAL,
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'close-annotation-delete-modal'
-    }
-  }
-});
-
-export const deleteAnnotation = (docId, annotationId) =>
-  (dispatch) => {
-    dispatch({
-      type: Constants.REQUEST_DELETE_ANNOTATION,
-      payload: {
-        annotationId
-      },
-      meta: {
-        analytics: {
-          category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-          action: 'request-delete-annotation'
-        }
-      }
-    });
-
-    ApiUtil.delete(`/document/${docId}/annotation/${annotationId}`, {}, ENDPOINT_NAMES.ANNOTATION).
-      then(
-        () => dispatch({
-          type: Constants.REQUEST_DELETE_ANNOTATION_SUCCESS,
-          payload: {
-            annotationId
-          }
-        }),
-        () => dispatch({
-          type: Constants.REQUEST_DELETE_ANNOTATION_FAILURE,
-          payload: {
-            annotationId
-          }
-        })
-      );
-  };
-
-export const cancelEditAnnotation = (annotationId) => ({
-  type: Constants.CANCEL_EDIT_ANNOTATION,
-  payload: {
-    annotationId
-  },
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'cancel-edit-annotation'
-    }
-  }
-});
-export const updateAnnotationContent = (content, annotationId) => ({
-  type: Constants.UPDATE_ANNOTATION_CONTENT,
-  payload: {
-    annotationId,
-    content
-  },
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'edit-annotation-content-locally',
-      debounceMs: 500
-    }
-  }
-});
-export const updateNewAnnotationContent = (content) => ({
-  type: Constants.UPDATE_NEW_ANNOTATION_CONTENT,
-  payload: {
-    content
-  }
+export const onScrollToComment = (scrollToComment) => ({
+  type: Constants.SCROLL_TO_COMMENT,
+  payload: { scrollToComment }
 });
 
 export const jumpToPage = (pageNumber, docId) => ({
@@ -130,87 +34,6 @@ export const jumpToPage = (pageNumber, docId) => ({
 export const resetJumpToPage = () => ({
   type: Constants.RESET_JUMP_TO_PAGE
 });
-
-export const requestEditAnnotation = (annotation) => (dispatch) => {
-  // If the user removed all text content in the annotation, ask them if they're
-  // intending to delete it.
-  if (!annotation.comment) {
-    dispatch(openAnnotationDeleteModal(annotation.id, 'open-by-deleting-all-annotation-content'));
-
-    return;
-  }
-
-  dispatch({
-    type: Constants.REQUEST_EDIT_ANNOTATION,
-    payload: {
-      annotationId: annotation.id
-    },
-    meta: {
-      analytics: {
-        category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-        action: 'request-edit-annotation'
-      }
-    }
-  });
-
-  const data = ApiUtil.convertToSnakeCase({ annotation });
-
-  ApiUtil.patch(`/document/${annotation.documentId}/annotation/${annotation.id}`, { data }, ENDPOINT_NAMES.ANNOTATION).
-    then(
-      () => dispatch({
-        type: Constants.REQUEST_EDIT_ANNOTATION_SUCCESS,
-        payload: {
-          annotationId: annotation.id
-        }
-      }),
-      () => dispatch({
-        type: Constants.REQUEST_EDIT_ANNOTATION_FAILURE,
-        payload: {
-          annotationId: annotation.id
-        }
-      })
-    );
-};
-
-export const createAnnotation = (annotation) => (dispatch) => {
-  const temporaryId = uuid.v4();
-
-  dispatch({
-    type: Constants.REQUEST_CREATE_ANNOTATION,
-    payload: {
-      annotation: {
-        ...annotation,
-        id: temporaryId
-      }
-    }
-  });
-
-  const data = ApiUtil.convertToSnakeCase({ annotation });
-
-  ApiUtil.post(`/document/${annotation.documentId}/annotation`, { data }, ENDPOINT_NAMES.ANNOTATION).
-    then(
-      (response) => {
-        const responseObject = JSON.parse(response.text);
-
-        dispatch({
-          type: Constants.REQUEST_CREATE_ANNOTATION_SUCCESS,
-          payload: {
-            annotation: {
-              ...annotation,
-              ...responseObject
-            },
-            annotationTemporaryId: temporaryId
-          }
-        });
-      },
-      () => dispatch({
-        type: Constants.REQUEST_CREATE_ANNOTATION_FAILURE,
-        payload: {
-          annotationTemporaryId: temporaryId
-        }
-      })
-    );
-};
 
 export const handleSetLastRead = (docId) => ({
   type: Constants.LAST_READ_DOCUMENT,
@@ -378,3 +201,97 @@ export const togglePdfSidebar = () => ({
     }
   }
 });
+
+export const setUpPdfPage = (file, pageIndex, page) => ({
+  type: Constants.SET_UP_PDF_PAGE,
+  payload: {
+    file,
+    pageIndex,
+    page
+  }
+});
+
+export const clearPdfPage = (file, pageIndex, page) => ({
+  type: Constants.CLEAR_PDF_PAGE,
+  payload: {
+    file,
+    pageIndex,
+    page
+  }
+});
+
+export const clearPdfDocument = (file, pageIndex, doc) => ({
+  type: Constants.CLEAR_PDF_DOCUMENT,
+  payload: {
+    file,
+    pageIndex,
+    doc
+  }
+});
+
+export const setPdfDocument = (file, doc) => ({
+  type: Constants.SET_PDF_DOCUMENT,
+  payload: {
+    file,
+    doc
+  }
+});
+
+export const rotateDocument = (docId) => ({
+  type: Constants.ROTATE_PDF_DOCUMENT,
+  payload: {
+    docId
+  }
+});
+
+export const getDocumentText = (pdfDocument, file) =>
+  (dispatch) => {
+    const getTextForPage = (index) => {
+      return pdfDocument.getPage(index + 1).then((page) => {
+        return page.getTextContent();
+      });
+    };
+    const getTextPromises = _.range(pdfDocument.pdfInfo.numPages).map((index) => getTextForPage(index));
+
+    Promise.all(getTextPromises).then((pages) => {
+      const textObject = pages.reduce((acc, page, pageIndex) => {
+        // PDFJS textObjects have an array of items. Each item has a str.
+        // concatenating all of these gets us to the page text.
+        const concatenated = page.items.map((row) => row.str).join(' ');
+
+        return {
+          ...acc,
+          [`${file}-${pageIndex}`]: {
+            id: `${file}-${pageIndex}`,
+            file,
+            text: concatenated,
+            pageIndex
+          }
+        };
+      }, {});
+
+      dispatch({
+        type: Constants.GET_DOCUMENT_TEXT,
+        payload: {
+          textObject
+        }
+      });
+    });
+  }
+;
+
+export const updateSearchIndex = (increment) => ({
+  type: Constants.UPDATE_SEARCH_INDEX,
+  payload: {
+    increment
+  }
+});
+
+export const zeroSearchIndex = () => ({
+  type: Constants.ZERO_SEARCH_INDEX
+});
+
+export const searchText = (searchTerm) => (dispatch) => {
+  dispatch(zeroSearchIndex());
+  dispatch(createSearchAction('extractedText')(searchTerm));
+};
